move Logo to the reusableComponents folder
