+++ anywhere-fitness/src/ReusableComponents/Logo.js
